requirements.txt
Django==4.1.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2==2.9.5
python-decouple==3.7